requirements.txt
alembic==1.14.0
bcrypt==4.2.0
blinker==1.9.0
cachelib==0.13.0
cffi==1.17.1
click==8.1.7
cryptography==43.0.3
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.8
MarkupSafe==3.0.2
msgspec==0.18.6
mysql-connector-python==9.1.0
mysqlclient==2.2.7
packaging==24.2
psycopg2-binary==2.9.10
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==1.0.1
redis==5.2.1
SQLAlchemy==2.0.36
sqlparse==0.5.1
termcolor==2.5.0
typing_extensions==4.12.2
Werkzeug==3.1.3
